Extract `CoreDataFeedStore` helper methods into their own file
